requirements.txt
alembic==1.9.2
bcrypt==4.0.1
cachetools==5.3.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
deepdiff==5.7.0
dnspython==2.3.0
docutils==0.18.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fpdf==1.7.2
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsondiff==1.3.1
Mako==1.2.4
MarkupSafe==2.1.1
mysql-connector-python==8.1.0
oauthlib==3.2.2
openpyxl==3.0.9
ordered-set==4.0.2
pdfkit==1.0.0
Pillow==10.0.1
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.11.2
pyparsing==3.1.1
pypng==0.20220715.0
python-dotenv==0.19.2
qrcode==7.4.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.17.6
six==1.16.0
snipeitpyapi==1.6
SQLAlchemy==1.4.46
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.2.2
wkhtmltopdf==0.2
WTForms==3.0.1
qrcode~=7.4.2